requirements.txt
# =============================================================
# Opsra — backend/requirements.txt
# Aligned with Technical Spec v1.0 confirmed tech stack
# =============================================================

# ── Core Framework ───────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==21.2.0
python-multipart==0.0.9       # File uploads — ticket attachments

# ── Configuration & Environment ──────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1       # Spec Section 2.2 — 16 env vars via BaseSettings
pydantic[email]==2.7.1         # Email validation on user/lead models
python-dotenv==1.0.0

# ── Database — Supabase ──────────────────────────────────────
supabase==2.4.6                # Supabase Python client (REST + Auth + Storage + Realtime)
postgrest==0.16.8              # Underlying PostgREST client used by supabase-py

# ── Authentication & Security ────────────────────────────────
python-jose[cryptography]==3.3.0   # JWT decode/verify for Supabase JWTs
passlib[bcrypt]==1.7.4             # Password hashing utilities
slowapi==0.1.9                     # Rate limiting — Spec Section 11.4 (login, AI endpoints)

# ── AI — Anthropic Claude ────────────────────────────────────
anthropic==0.26.1              # Sonnet + Haiku — Spec Section 1 & Section 8/12

# ── Background Jobs — Celery + Redis ────────────────────────
celery[redis]==5.3.6           # Task queue — Spec Section 1 (12 scheduled jobs)
redis==5.0.4                   # Redis client — Upstash Redis via rediss:// TLS

# ── HTTP Client ──────────────────────────────────────────────
httpx==0.27.0                  # Async HTTP — Meta Cloud API calls, internal service calls

# ── Email — Resend ───────────────────────────────────────────
resend==2.0.0                  # Transactional email — Spec Section 1

# ── Webhook Security ─────────────────────────────────────────
cryptography==42.0.5           # HMAC signature verification — Meta & payment webhooks

# ── Utilities ────────────────────────────────────────────────
python-dateutil==2.9.0         # Date arithmetic — SLA timers, renewal reminders
pytz==2024.1                   # Timezone handling — org timezone field
bleach==6.1.0                  # Input sanitisation — ticket/message content

# =============================================================
# Testing Dependencies
# Per Build Status doc — Testing Stack section
# =============================================================
pytest==7.4.0
pytest-asyncio==0.21.0         # Async test support for FastAPI async routes
httpx==0.27.0                  # AsyncClient for route integration tests
factory-boy==3.3.0             # Test data factories — org, user, lead, ticket
respx==0.20.2                  # Mock httpx calls — Anthropic API + Meta API
pytest-cov==4.1.0              # Coverage reporting — target 80%+ per Build Status

